Allow async transforms
