+++ lib/index.js
@@ -1,79 +1,61 @@
 const debug = require('debug')('metalsmith-in-place');
 const match = require('multimatch');
 const isUtf8 = require('is-utf8');
-const jstransformer = require('jstransformer');
-const toTransformer = require('inputformat-to-jstransformer');
-
-/**
- * Gets jstransformer for an extension, and caches them
- */
-
-const cache = {};
-
-function getTransformer(ext) {
-  if (ext in cache) {
-    return cache[ext];
-  }
-
-  const transformer = toTransformer(ext);
-  cache[ext] = transformer ? jstransformer(transformer) : false;
-
-  return cache[ext];
-}
+const { getTransformer } = require('./util');
 
 /**
  * Engine, renders file contents with all available transformers
  */
 
 function render({ filename, files, metadata, engineOptions }) {
-  return new Promise((resolve, reject) => {
-    const [base, ...extensions] = filename.split('.');
-    const file = files[filename];
+  const [base, ...extensions] = filename.split('.');
+  const file = files[filename];
 
-    debug(`rendering ${filename}`);
+  debug(`rendering ${filename}`);
 
-    // Stringify file contents
-    file.contents = file.contents.toString();
+  const ext = extensions.pop();
+  const transform = getTransformer(ext);
+  const locals = Object.assign({}, metadata, file);
 
-    // Go through all extensions
-    const extLength = extensions.length;
-    for (let i = 0; i < extLength; i += 1) {
-      const ext = extensions.pop();
-      const transform = getTransformer(ext);
-      const locals = Object.assign({}, metadata, file);
+  // Stop if the current extension can't be transformed
+  if (!transform) {
+    debug(`no transformer available for ${ext} extension for ${filename}`);
+    extensions.push(ext);
+    return Promise.resolve();
+  }
 
-      // Stop if the current extension can't be transformed
-      if (!transform) {
-        debug(`no transformer available for ${ext} extension for ${filename}`);
-        extensions.push(ext);
-        break;
-      }
+  // Stringify file contents
+  file.contents = file.contents.toString();
 
-      // If this is the last extension, replace it with a new one
-      if (extensions.length === 0) {
-        debug(`last extension reached, replacing last extension with ${transform.outputFormat}`);
-        extensions.push(transform.outputFormat);
-      }
+  // If this is the last extension, replace it with a new one
+  if (extensions.length === 0) {
+    debug(`last extension reached, replacing last extension with ${transform.outputFormat}`);
+    extensions.push(transform.outputFormat);
+  }
 
-      // Transform the contents
-      debug(`rendering ${ext} extension for ${filename}`);
-      try {
-        file.contents = transform.render(file.contents, engineOptions, locals).body;
-      } catch (err) {
-        err.message = `${filename}: ${err.message}`;
-        return reject(err);
-      }
-    }
+  // Transform the contents
+  debug(`rendering ${ext} extension for ${filename}`);
 
-    // Store results and delete old file
-    file.contents = Buffer.from(file.contents);
-    const newName = [base, ...extensions].join('.');
-    files[newName] = file; // eslint-disable-line no-param-reassign
-    delete files[filename]; // eslint-disable-line no-param-reassign
+  return transform
+    .renderAsync(file.contents, engineOptions, locals)
+    .then(rendered => {
+      // Update contents and delete old file
+      file.contents = Buffer.from(rendered.body);
+      delete files[filename]; // eslint-disable-line no-param-reassign
 
-    debug(`done rendering ${filename}, renamed to ${newName}`);
-    return resolve();
-  });
+      // Update files with the newly rendered file
+      const newName = [base, ...extensions].join('.');
+      files[newName] = file; // eslint-disable-line no-param-reassign
+
+      debug(`done rendering ${filename}, renamed to ${newName}`);
+
+      // Keep rendering until there are no applicable transformers left
+      return render({ filename: newName, files, metadata, engineOptions });
+    })
+    .catch(err => {
+      err.message = `${filename}: ${err.message}`; // eslint-disable-line no-param-reassign
+      throw err;
+    });
 }
 
 /**
