+++ lib/util.js
@@ -1,0 +1,23 @@
+const jstransformer = require('jstransformer');
+const toTransformer = require('inputformat-to-jstransformer');
+
+/**
+ * Gets jstransformer for an extension, and caches them
+ */
+
+const cache = {};
+
+function getTransformer(ext) {
+  if (ext in cache) {
+    return cache[ext];
+  }
+
+  const transformer = toTransformer(ext);
+  cache[ext] = transformer ? jstransformer(transformer) : false;
+
+  return cache[ext];
+}
+
+module.exports = {
+  getTransformer
+};
